Fixed poolAddress and PriceInETH, Test Borrow GHO, Test Repay GHO, Test position Status
Re-fixed the issue we had when computing the pool address in openPosition as well as the price in ETH when the token is ETH itself (we use to return 1 but since it has 18 decimals we had to return 10**18).

Implemented some tests for the GHO borrowing and repaying mechanism:
- testBorrowAndRepayGHO_wrongUserBorrow,
- testBorrowAndRepayGHO_wrongUserRepay,
- testBorrowAndRepayGHO_checkDebtEvolution,
- testBorrowAndRepayGHO_borrow0GHO,
- testBorrowAndRepayGHO_repay0GHO,
- testBorrowAndRepayGHO_repayMoreThanDebt

Implemented some tests regarding the status of a position:
- testPositionStatus_closeByOwner,
- testPositionStatus_closeByOwnerDebtNotRepaid,
- testPositionStatus_closeByOwnerDebtRepaid,
- testPositionStatus_notExistent

WARNING: I don't know why but when we are creating positions for factice users the amount of tokens in those positions are 0.